pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "layerbot"
version = "0.1.0"
description = "A bot for monitoring Layer bridge deposits"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "click>=8.0.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0"
]

[project.scripts]
layerbot = "layerbot.cli:cli"

[tool.hatch.build.targets.wheel]
packages = ["src/layerbot"] 